chore: rename components folder to _components for routing exclusion
