+++ trip-talk/src/app/_components/AddressModal/AddressModal.tsx
@@ -1,0 +1,3 @@
+export default function AddressModal() {
+  return;
+}
