+++ trip-talk/src/app/_components/BannerSlider/BannerSlider.tsx
