+++ trip-talk/src/app/_components/BoardList/BoardList.tsx
