+++ trip-talk/src/app/_components/BoardsDetail/BoardsDetail.tsx
