+++ trip-talk/src/app/_components/BoardsWrite/BoardsWrite.tsx
