+++ trip-talk/src/app/_components/CustomImage/CustomImage.tsx
