+++ trip-talk/src/app/_components/LikeButton/LikeButton.tsx
