+++ trip-talk/src/app/_components/StarRating/StarRating.tsx
