+++ trip-talk/src/app/_components/TextArea/TextArea.tsx
