+++ trip-talk/src/app/boards/[boardId]/page.tsx
@@ -1,6 +1,6 @@
-import BoardsDetail from "../../components/BoardsDetail/BoardsDetail";
-import CommentList from "../../components/CommentList/CommentList";
-import CommentWrite from "../../components/CommentWrite/CommentWrite";
+import BoardsDetail from "../../_components/BoardsDetail/BoardsDetail";
+import CommentList from "../../_components/CommentList/CommentList";
+import CommentWrite from "../../_components/CommentWrite/CommentWrite";
 
 export default function BoardsDetailPage() {
   return (
