--- trip-talk/src/app/boards/styles.module.css
@@ -1,31 +1,0 @@
-.post_contain {
-  @apply flex flex-col items-start gap-2 self-stretch w-full rounded-2xl bg-white py-6 px-12;
-  min-width: 680px;
-  box-shadow: 0px 0px 20px 0px rgba(0, 0, 0, 0.08);
-}
-
-.posts {
-  @apply w-full flex flex-col gap-2;
-}
-
-.boards_header {
-  @apply flex items-center gap-2 self-stretch py-4 px-6 text-base font-medium text-charcoal;
-}
-
-.boards_header_number {
-  @apply flex w-16 items-center justify-center gap-2.5;
-}
-
-.boards_header_title {
-  @apply flex gap-2.5 items-center justify-start flex-1;
-}
-
-.boards_header_writer {
-  @apply flex justify-center items-center gap-2.5;
-  width: 100px;
-}
-
-.boards_header_createdAt {
-  @apply flex justify-center items-center gap-2.5;
-  width: 100px;
-}
